+++ src/controller/category.controller.ts
@@ -2,6 +2,7 @@ import { Response } from "express";
 import { IReqUser, IPaginatinationQuery } from "../utils/interface";
 import CategoryModel, { categoryDAO } from "../models/category.model";
 import response from "../utils/response";
+import { isValidObjectId } from "mongoose";
 
 export default {
   async create(req: IReqUser, res: Response) {
@@ -59,7 +60,14 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed find one a category");
+      }
+
       const result = await CategoryModel.findById(id);
+      if (!result) {
+        return response.notFound(res, "failed find one a category");
+      }
 
       response.success(res, result, "success find one category");
     } catch (error) {
@@ -70,6 +78,10 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed update a category");
+      }
+
       const result = await CategoryModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -83,6 +95,10 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed remove a category");
+      }
+
       const result = await CategoryModel.findByIdAndDelete(id, {
         new: true,
       });
